fastapi/requirements.txt
fastapi
sqlalchemy
pydantic
python-dotenv
uvicorn[standard]
psycopg2-binary